backend/requirements.txt
fastapi[all]
sqlmodel
python-dotenv